Changed AprRate properties from decimals to doubles

+++ backend.Data/Models/MortgageAccount.cs
@@ -11,7 +11,7 @@ namespace backendData.Models
     public string Type { get; set; }
     public DateTime StartDate { get; set; }
     public int TermInMonths { get; set; }
-    public decimal AprRate { get; set; }    
+    public double AprRate { get; set; }    
     public bool IsActive { get; set; }
     [Required] public Currency QuotedCurrency { get; set; }
     [Required] public User User { get; set; }
